refactor: split watchers.py into per-source modules
The narrative rewrite revealed three self-contained subsystems stapled
into one file. Split into a package so the file tree tells the story:

  watchers/
    __init__.py       — overview docstring, re-exports public API
    _event.py         — WatchEvent (shared vocabulary)
    dns_log.py        — DnsLogWatcher + query regex + domain refresh
    audit_log.py      — AuditLogWatcher
    nflog.py          — NflogWatcher + netlink constants + helpers
    domain_cache.py   — DomainCache (enrichment helper)

Public imports unchanged (from terok_shield.lib.watchers import ...).
Test patch paths updated to target the specific submodule.

No functional changes.
